fix(warnings): clean up actionable compiler warnings
- optimization.cpp: init int_value/float_value to silence
  -Wconditional-uninitialized (guarded false positives)
- debugger.cpp: cast byte through uintptr_t before void* to fix
  -Wint-to-void-pointer-cast (preserves hex output)
- common.cpp: mark HTTP/2 & HTTP/3 stack-balance pops [[maybe_unused]]
  (only used when OBJECK_HAS_NGHTTP2/NGTCP2 defined); drop unused
  prio_ret/cfg_ret return-code locals
- context.cpp: remove dead full_entry_name/full_entry_index locals

+++ core/debugger/debugger.cpp
@@ -794,7 +794,7 @@ void Runtime::Debugger::ProcessPrint(Print* print) {
 
         case BYTE_ARY_PARM:
           if(reference->GetIndices()) {
-            std::wcout << L"print: type=Byte, value=" << (void*)((unsigned char)reference->GetIntValue()) << std::endl;
+            std::wcout << L"print: type=Byte, value=" << (void*)(uintptr_t)((unsigned char)reference->GetIntValue()) << std::endl;
           }
           else {
             std::wcout << L"print: type=Byte[], value=" << reference->GetIntValue() << L"(" << (void*)reference->GetIntValue() << L")";
@@ -931,7 +931,7 @@ void Runtime::Debugger::ProcessPrint(Print* print) {
           else if(ref_klass && ref_klass->GetName() == L"System.ByteRef") {
             size_t* instance = (size_t*)reference->GetIntValue();
             if(instance) {
-              std::wcout << L"print: type=System.ByteHolder, value=" << (void*)((unsigned char)instance[0]) << std::endl;
+              std::wcout << L"print: type=System.ByteHolder, value=" << (void*)(uintptr_t)((unsigned char)instance[0]) << std::endl;
             }
             else {
               std::wcout << L"print: type=" << (ref_klass ? ref_klass->GetName() : L"System.Base") << L", value=" << (void*)reference->GetIntValue() << std::endl;
